more reorganisation and licenses
