+++ internal/graphicscommand/image_notjs.go
@@ -11,24 +11,19 @@ import (
 )
 
 func (i *Image) Dump(graphicsDriver graphicsdriver.Graphics, path string, blackbg bool, rect image.Rectangle) error {
-	data, err := i.DumpBytes(graphicsDriver, blackbg, rect)
-	if err != nil {
-		return err
-	}
-
-	// screen image
-	if data == nil {
+	// Screen image cannot be dumped.
+	if i.screen {
 		return nil
 	}
 
-	path = i.FormatPath(path)
+	path = i.DumpName(path)
 	f, err := os.Create(path)
 	if err != nil {
 		return err
 	}
 	defer f.Close()
 
-	if _, err := f.Write(data); err != nil {
+	if err := i.DumpTo(graphicsDriver, blackbg, rect, f); err != nil {
 		return err
 	}
